pyproject.toml
[project]
name = "finbot"
dynamic = ["version"]
dependencies = [
    "aiohttp==3.11.18",
    "alembic==1.16.1",
    "bcrypt==4.3.0",
    "celery==5.5.2",
    "cffi==1.17.1",
    "cryptography==45.0.2",
    "fastapi[standard]<1",
    "gspread==6.2.1",
    "gunicorn==23.0.0",
    "jsonschema==4.23.0",
    "krakenex==2.2.2",
    "oauth2client==4.1.3",
    "orjson==3.10.18",
    "pika==1.3.2",
    "plaid-python==32.0.0",
    "playwright==1.52.0",
    "price-parser==0.4.0",
    "psycopg2==2.9.10",
    "pycoingecko==3.2.0",
    "pydantic==2.11.5",
    "python-binance==1.0.29",
    "python-gnupg==0.5.4",
    "python-jose[cryptography]<4",
    "requests==2.32.3",
    "schedule==1.2.2",
    "spectree==1.2.10",
    "sqlalchemy==2.0.41",
    "twilio==9.6.1",
    "uvicorn<1",
    "yfinance==0.2.61",
]

[project.optional-dependencies]
dev = [
    "click==8.2.1",
    "mypy==1.15.0",
    "pika-stubs==0.1.3",
    "pytest==8.3.5",
    "ruff==0.11.11",
    "semver==3.0.4",
    "sqlalchemy-stubs==0.4",
    "types-requests==2.32.0.20250515",
    "watchdog[watchmedo]<7",
]

[tool.ruff]
exclude = [
    'migrations/',
    'venv/',
    'webapp/'
]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "Q", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.setuptools.dynamic]
version = {attr = "finbot._version.__version__"}
